+++ packages/common/src/content/_internal/computeProps.ts
@@ -1,9 +1,13 @@
 import { IRequestOptions } from "@esri/arcgis-rest-request";
 import { UserSession } from "@esri/arcgis-rest-auth";
 import { getItemThumbnailUrl } from "../../resources";
-import { IModel } from "../../types";
+import { IHubRequestOptions, IModel } from "../../types";
 import { getItemHomeUrl } from "../../urls/get-item-home-url";
-import { getContentEditUrl, getHubRelativeUrl } from "./internalContentUtils";
+import {
+  getAdditionalResources,
+  getContentEditUrl,
+  getHubRelativeUrl,
+} from "./internalContentUtils";
 import { IHubEditableContent } from "../../core/types/IHubEditableContent";
 import { getRelativeWorkspaceUrl } from "../../core/getRelativeWorkspaceUrl";
 import { isDiscussable } from "../../discussions";
@@ -12,6 +16,7 @@ import {
   ServiceCapabilities,
 } from "../hostedServiceUtils";
 import { computeBaseProps } from "../../core/_internal/computeBaseProps";
+import { getProp } from "../../objects";
 import { IHubEditableContentEnrichments } from "../../items/_enrichments";
 
 export function computeProps(
@@ -57,6 +62,20 @@ export function computeProps(
     content.serverExtractCapability = hasServiceCapability(
       ServiceCapabilities.EXTRACT,
       enrichments.server
+    );
+    const extractFormatsList: string = getProp(
+      enrichments,
+      "server.supportedExportFormats"
+    );
+    content.serverExtractFormats =
+      extractFormatsList && extractFormatsList.split(",");
+  }
+
+  if (enrichments.metadata) {
+    content.additionalResources = getAdditionalResources(
+      model.item,
+      enrichments.metadata,
+      requestOptions as IHubRequestOptions
     );
   }