+++ packages/common/CHANGELOG.md
@@ -1,3 +1,24 @@
+# @esri/hub-common [14.122.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.121.2...@esri/hub-common@14.122.0) (2024-05-01)
+
+
+### Features
+
+* **hub-common:** add new utils for a unified download flow ([#1482](https://github.com/Esri/hub.js/issues/1482)) ([7f45d61](https://github.com/Esri/hub.js/commit/7f45d61bec8c8b79227834df6ac02ea4ae7f4328))
+
+## @esri/hub-common [14.121.2](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.121.1...@esri/hub-common@14.121.2) (2024-04-30)
+
+
+### Bug Fixes
+
+* **hub-common:** add startDate, startTime, endDate, endTime to create… ([#1480](https://github.com/Esri/hub.js/issues/1480)) ([74b4634](https://github.com/Esri/hub.js/commit/74b463485c9583281ed7a919e8974153bcd0ecd2))
+
+## @esri/hub-common [14.121.1](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.121.0...@esri/hub-common@14.121.1) (2024-04-29)
+
+
+### Bug Fixes
+
+* export IWithFollowers ([#1489](https://github.com/Esri/hub.js/issues/1489)) ([6455b42](https://github.com/Esri/hub.js/commit/6455b421425995f71b3f8578114deea0033b680c))
+
 # @esri/hub-common [14.121.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.120.0...@esri/hub-common@14.121.0) (2024-04-29)